+++ src/CanadianCuisine/Behaviours/ActionSpawnOnFirstUse.cs
@@ -5,11 +5,10 @@ public class ActionSpawnOnFirstUse : Action_Spawn
     public override void RunAction()
     {
         var uses = item.GetData<OptionableIntItemData>(DataEntryKey.ItemUses);
-        
-        if (!uses.HasData || uses.Value == item.totalUses)
+
+        if (!uses.HasData || uses.Value == (item.totalUses - 1))
         {
             base.RunAction();
         }
-        
     }
 }